refactor: rename robot_data_spec_utils to embodiment_description_utils
