+++ neuracore/ml/utils/training_config.py
@@ -14,7 +14,7 @@ import neuracore as nc
 from neuracore.api.training import _get_algorithms, get_algorithm
 from neuracore.core.const import DEFAULT_CACHE_DIR
 from neuracore.core.data.dataset import Dataset
-from neuracore.core.utils.robot_data_spec_utils import (
+from neuracore.core.utils.embodiment_description_utils import (
     convert_cross_embodiment_description_names_to_ids,
     is_robot_id,
 )
